// src/types/enums.ts
// Enum for client age groups
export enum AGEGROUP {
  MINOR = "minor",
  ADULT = "adult",
  SENIOR = "senior",
}

// Enum for service types offered
export enum SERVICES {
  SHOWER = "SHOWER",
  LAUNDRY = "LAUNDRY",
  MNGMT = "MNGMT",
}

// Enum for user/staff roles (optional for access control)
export enum USERROLE {
  ADMIN = "admin",
  CASE_MANAGER = "case-manager",
  STAFF = "staff",
}